+++ test/otel/exporter/chdb_test.clj
@@ -20,6 +20,7 @@
             [otel.exporter.chdb-typed-query-test :as typed-query-test]
             [otel.exporter.chdb-typed-query-native-test :as typed-query-native-test]
             [otel.exporter.chdb-typed-log-test :as typed-log-test]
+            [otel.exporter.chdb-typed-gauge-test :as typed-gauge-test]
             [otel.exporter.chdb-typed-log-explorer-test :as typed-log-explorer-test]
             [otel.exporter.chdb-typed-log-socket-native-test :as typed-log-socket-test]
             [otel.exporter.chdb-benchmark :as benchmark]
@@ -782,6 +783,7 @@
   (registry-installer-test/run check)
   (typed-query-test/run check)
   (typed-log-test/run check)
+  (typed-gauge-test/run check)
   (typed-log-explorer-test/run check)
   (typed-query-native-test/run check)
   (typed-log-socket-test/-main)